Canned routine for AHP

ahp is a canned routine that returns a data.frame of individuals' priorities (based on ahp.indpref), number of missing values, consistency ratios, and top three pairs of inconsistent pairwise comparisons. An optional argument agg allows the user to generate a data.frame of aggregated priorities alongside individual priorities.

Usage

ahp(
  df,
  atts,
  negconvert = FALSE,
  reciprocal = TRUE,
  method = "eigen",
  aggmethod = "geometric",
  qt = 0,
  censorcr = Inf,
  agg = FALSE,
  ID = NULL,
  col = NULL,
  suppress = "none"
)

Arguments

df

a dataframe, each row corresponding to one decision-maker, with columns ordered according to atts.

atts

a list of attributes in the correct order

negconvert

logical, whether to convert all positive values to negative. In the pairwise comparison A_B, if -6 denotes A is more important than B by 6 units, set negconvert = TRUE.

reciprocal

logical, whether to convert negative values (after negconvert) to its reciprocal. If the comparison A_B where B is more important than A was already entered in its reciprocal (e.g. 1/6), choose reciprocal = FALSE. When reciprocal = FALSE, do not set negconvert = TRUE.

method

if method = "eigen", the individual priority weights are computed using the Dominant Eigenvalues method described in Saaty (2003) . Otherwise, then the priorities are computed based on the averages of normalized values. Basically it normalizes the matrices so that all of the columns add up to 1, and then computes the averages of the row as the priority weights of each attribute. Three modes of finding the averages are available: arithmetic: the arithmetic mean; geometric: the geometric mean (the default); rootmean: the square root of the sum of the squared value.

aggmethod

how to aggregate the individual priorities. By default aggmethod = method. Apart from the methods offered in method, aggmethod also permits three other options: tmean computes the trimmed arithmetic mean, tgmean computes the trimmed geometric mean (both with quantiles trimmed based on qt), and sd computes the standard deviation from the arithmetic mean. If method = "eigen" and aggmethod is not specified, aggmethod defaults to "geometric".

qt

specifies the quantile which the top and bottom priority weights are trimmed. Used only if aggmethod = 'tmean' or aggmethod = 'tgmean'. For example, qt = 0.25 specifies that the aggregation is the arithmetic mean of the values from the 25 to 75 percentile. By default qt = 0.

censorcr

the threshold of censoring observations. All observations with a consistency ratio higher than the specified CR will be dropped from the routine. By default, censorcr = Inf, i.e. drops no observations.

agg

by default, agg = FALSE, which suppresses the output of Aggregated priorities. If agg = TRUE, the output will be a list where $AggPref contains the Aggregated priorities and Standard Deviations.

ID

the column name, or a vector of column names, of variables in df, which are used to indicate specific observations. For example, if a survey dataset comes with an identifier/ multiple identified for each observation, use this so that the output dataframe would contain this identifier.

col

a list of column names which contains the columns for the pairwise comparison matrix in chronological order, as specified in atts. For example, if atts is c("A","B","C"), then col would be c("A_B","A_C","B_C"), with "A_B" being a pairwise comparison of attribute A and B. col = NULL by default, indicating that the dataframe specified already contained the pairwise comparisons in correct order and nothing else except of what is inside ID.

suppress

Suppresses the supplementary indices. Can take the values of "all" (suppresses all indices), "pwerror" (suppresses pwerror), or "cr" (suppresses consistency ratio).

Value

If agg = TRUE, returns a list of the results with two elements, indpref and aggpref. $indpref contains individual preference weight (same as ahp.indpref), consistency ratios, and the top three most inconsistent pairwise comparisons (same as ahp.pwerror). $aggpref contains the Aggregated Priorities and Standard Deviations. Otherwise (i.e. default setting), outputs a data.frame with only $indpref.

References

Saaty TL (2004). “Decision making — the Analytic Hierarchy and Network Processes (AHP/ANP).” Journal of Systems Science and Systems Engineering, 13(1), 1–35. ISSN 1861-9576. doi:10.1007/s11518-006-0151-5 .

Author

Frankie Cho

Examples

data(city200)
ahp(df = city200, atts = c('cult', 'fam', 'house', 'jobs', 'trans'), negconvert = TRUE)
#>           cult       fam      house      jobs      trans          CR       top1
#> 1   0.17094660 0.4587181 0.08547330 0.2507636 0.03409845 0.061358609   fam_jobs
#> 2   0.22910090 0.3935620 0.08292558 0.2531962 0.04121537 0.029678314 cult_house
#> 3   0.15400448 0.4921905 0.08239372 0.2213908 0.05002052 0.063388311  fam_trans
#> 4   0.12424954 0.4634863 0.06162027 0.3159930 0.03465092 0.093246803   cult_fam
#> 5   0.15216759 0.3556904 0.07239889 0.3748108 0.04493236 0.106226128  cult_jobs
#> 6   0.15365597 0.4738939 0.07106456 0.2516808 0.04970479 0.107590264  fam_trans
#> 7   0.21361837 0.4926880 0.05269318 0.2060430 0.03495745 0.213740701  cult_jobs
#> 8   0.14337912 0.5022318 0.06034660 0.2569918 0.03705070 0.061735365  cult_jobs
#> 9   0.16361589 0.4270270 0.07698011 0.2915773 0.04079969 0.038342589  fam_trans
#> 10  0.12877331 0.5194998 0.06354422 0.2489541 0.03922858 0.119328991  fam_trans
#> 11  0.15702250 0.4150990 0.07247617 0.3022023 0.05320010 0.136351544  fam_trans
#> 12  0.19484502 0.4062611 0.07771452 0.2620788 0.05910049 0.238768605  fam_trans
#> 13  0.19214133 0.4268636 0.07250191 0.2695353 0.03895785 0.036267286  fam_trans
#> 14  0.13016766 0.4667971 0.06212555 0.2996915 0.04121821 0.038479583  cult_jobs
#> 15  0.16457993 0.4903790 0.06919979 0.2430168 0.03282453 0.070583192   cult_fam
#> 16  0.14272403 0.4864032 0.07985419 0.2376762 0.05334235 0.090951139  fam_trans
#> 17  0.13586720 0.4972537 0.06728171 0.2584289 0.04116847 0.121043385   cult_fam
#> 18  0.12890969 0.4696133 0.06814983 0.2981436 0.03518358 0.058082463  cult_jobs
#> 19  0.14174902 0.5144712 0.06227759 0.2467923 0.03470994 0.092924026   cult_fam
#> 20  0.13503941 0.4938626 0.07518477 0.2392947 0.05661859 0.064562823  fam_trans
#> 21  0.18494119 0.4683249 0.07078184 0.2351782 0.04077392 0.071010525   fam_jobs
#> 22  0.18621850 0.4588558 0.06979270 0.2416068 0.04352625 0.071528763  fam_trans
#> 23  0.19932890 0.5376039 0.05179894 0.1703539 0.04091437 0.209664724  cult_jobs
#> 24  0.19839818 0.3528194 0.06691437 0.3388022 0.04306589 0.122468650  cult_jobs
#> 25  0.18545399 0.4357040 0.09026245 0.2228880 0.06569162 0.203722340  fam_trans
#> 26  0.14897235 0.4280356 0.09137098 0.2863602 0.04526084 0.110459779  fam_trans
#> 27  0.17289656 0.4384908 0.07231134 0.2756411 0.04066021 0.100961643  fam_trans
#> 28  0.16511099 0.4150944 0.07420459 0.3053381 0.04025188 0.044414042  cult_jobs
#> 29  0.10436683 0.5343955 0.06663420 0.2475962 0.04700729 0.160878654  fam_trans
#> 30  0.15166388 0.4253701 0.06901181 0.3154191 0.03853507 0.183139506   fam_jobs
#> 31  0.15290870 0.3841884 0.06267879 0.3682729 0.03195118 0.187920116   fam_jobs
#> 32  0.15659488 0.5065770 0.06641977 0.2348861 0.03552221 0.083195119   cult_fam
#> 33  0.13316751 0.4784325 0.06703099 0.2808590 0.04051000 0.075105744  fam_trans
#> 34  0.17191340 0.4511853 0.08026985 0.2603221 0.03630929 0.020024629 cult_house
#> 35  0.23530601 0.4742200 0.06002960 0.1856040 0.04484042 0.147699219  fam_trans
#> 36  0.10891359 0.4798659 0.07680004 0.2989572 0.03546320 0.070009799   cult_fam
#> 37  0.14379816 0.3553386 0.07477545 0.3834565 0.04263129 0.095982691   fam_jobs
#> 38  0.14486747 0.4813407 0.07684265 0.2629940 0.03395521 0.046376412 house_jobs
#> 39  0.22720575 0.3182677 0.07492863 0.3355041 0.04409375 0.026248174  fam_trans
#> 40  0.18035592 0.4466319 0.06741793 0.2605365 0.04505770 0.031794763  fam_trans
#> 41  0.18072194 0.3770409 0.07211986 0.3205327 0.04958460 0.133082581   fam_jobs
#> 42  0.19442581 0.5236417 0.05712660 0.1916790 0.03312685 0.101034537  cult_jobs
#> 43  0.13823759 0.4779618 0.06428916 0.2865851 0.03292641 0.046670677  cult_jobs
#> 44  0.11990104 0.4961348 0.06044860 0.2777800 0.04573548 0.249488508  fam_trans
#> 45  0.14171951 0.5144328 0.06090635 0.2460329 0.03690840 0.104985409  fam_trans
#> 46  0.27923833 0.3515989 0.07614933 0.2457271 0.04728635 0.055496664  fam_trans
#> 47  0.19961004 0.3431485 0.07999374 0.3360627 0.04118506 0.031937753  fam_trans
#> 48  0.13277872 0.5015437 0.06631841 0.2635715 0.03578765 0.074873282  cult_jobs
#> 49  0.25567557 0.4710978 0.05774544 0.1837438 0.03173741 0.103260200   fam_jobs
#> 50  0.17000053 0.4493720 0.07349909 0.2627970 0.04433143 0.076104146 cult_trans
#> 51  0.13331937 0.5037946 0.06134714 0.2690979 0.03244098 0.048427116  cult_jobs
#> 52  0.12172362 0.4900131 0.07015001 0.2742772 0.04383607 0.046627168 cult_trans
#> 53  0.13783774 0.4614154 0.06887571 0.2954564 0.03641478 0.099756391  cult_jobs
#> 54  0.16110965 0.5503193 0.05274605 0.2004664 0.03535865 0.080913909   cult_fam
#> 55  0.15332774 0.4419949 0.07734913 0.2846016 0.04272669 0.085876718  fam_trans
#> 56  0.10446338 0.5276956 0.06802125 0.2630496 0.03677026 0.061202622   cult_fam
#> 57  0.15206280 0.4652867 0.06939431 0.2709053 0.04235094 0.084693426  fam_trans
#> 58  0.19941873 0.4172036 0.07923022 0.2613200 0.04282739 0.088407306  fam_trans
#> 59  0.13047718 0.4486208 0.06290785 0.3234364 0.03455778 0.077458293  cult_jobs
#> 60  0.13726905 0.4120681 0.07078224 0.3239255 0.05595515 0.150118275  fam_trans
#> 61  0.14141207 0.3870909 0.10286016 0.3180711 0.05056570 0.042794076  fam_trans
#> 62  0.13759239 0.4328046 0.06109356 0.3332953 0.03521412 0.154374565   fam_jobs
#> 63  0.10849861 0.4834296 0.06802352 0.3070491 0.03299917 0.067671611  cult_jobs
#> 64  0.20471441 0.4283026 0.06630874 0.2618415 0.03883276 0.064144942 cult_house
#> 65  0.14008117 0.5183553 0.06462340 0.2437453 0.03319487 0.081833743   cult_fam
#> 66  0.15144250 0.5142276 0.06845022 0.2196245 0.04625525 0.090600905  fam_trans
#> 67  0.13706986 0.4927880 0.07387374 0.2510442 0.04522423 0.127877729  fam_trans
#> 68  0.12846293 0.5051158 0.06922509 0.2546816 0.04251455 0.095472359   cult_fam
#> 69  0.13475919 0.4850243 0.07188846 0.2629399 0.04538811 0.132920109  fam_trans
#> 70  0.11600429 0.4721881 0.07792130 0.2900496 0.04383673 0.084074539  cult_jobs
#> 71  0.13212977 0.4328588 0.06911904 0.3250852 0.04080719 0.063177885  fam_trans
#> 72  0.11877000 0.5104478 0.07143246 0.2621187 0.03723099 0.023985432   cult_fam
#> 73  0.23392576 0.4172835 0.06546533 0.2455678 0.03775757 0.106478426  cult_jobs
#> 74  0.10643845 0.3852796 0.07018755 0.4019494 0.03614505 0.128437131  cult_jobs
#> 75  0.15912317 0.4913838 0.07770426 0.2376506 0.03413814 0.036202974   cult_fam
#> 76  0.15949069 0.5171183 0.07466063 0.2064552 0.04227515 0.109946172  fam_trans
#> 77  0.17489444 0.3974014 0.07803729 0.3154891 0.03417776 0.129504947   fam_jobs
#> 78  0.20452106 0.4880581 0.06752525 0.2033243 0.03657126 0.065569152  cult_jobs
#> 79  0.15665639 0.4621967 0.07842151 0.2616708 0.04105460 0.085181531  fam_trans
#> 80  0.12454237 0.4677734 0.06185714 0.3025820 0.04324510 0.053589390  cult_jobs
#> 81  0.19544761 0.3710763 0.07889933 0.3194113 0.03516547 0.032549198 cult_house
#> 82  0.17485872 0.3594333 0.06119124 0.3740339 0.03048280 0.185076362   fam_jobs
#> 83  0.12141446 0.4174045 0.07430552 0.3442454 0.04263011 0.071260192  cult_jobs
#> 84  0.16091177 0.4773699 0.06640949 0.2462238 0.04908504 0.096470970 cult_trans
#> 85  0.11464303 0.4918300 0.07750067 0.2761527 0.03987369 0.115944371  cult_jobs
#> 86  0.14454910 0.4694915 0.06753044 0.2798321 0.03859692 0.103167626  fam_trans
#> 87  0.13712643 0.3453075 0.06599252 0.4162521 0.03532149 0.101429657   fam_jobs
#> 88  0.14211431 0.4051082 0.06943007 0.3413200 0.04202739 0.071065777  fam_trans
#> 89  0.12920638 0.5286750 0.06700711 0.2339455 0.04116603 0.053085253  fam_trans
#> 90  0.15647292 0.4828176 0.06553624 0.2613061 0.03386716 0.036446340  fam_trans
#> 91  0.27144874 0.3417462 0.07526555 0.2792685 0.03227102 0.174606380  cult_jobs
#> 92  0.15890802 0.4903858 0.07425969 0.2372065 0.03923998 0.062995871  fam_trans
#> 93  0.13814121 0.4556754 0.06992901 0.3025411 0.03371324 0.110471976  cult_jobs
#> 94  0.18510140 0.3571961 0.06887929 0.3520718 0.03675145 0.023703938  cult_jobs
#> 95  0.21685317 0.4260865 0.07849586 0.2399713 0.03859317 0.014622927 cult_house
#> 96  0.13639014 0.4312784 0.08839278 0.3045112 0.03942743 0.085630588   cult_fam
#> 97  0.13603466 0.3750731 0.08669576 0.3640091 0.03818743 0.071560167  cult_jobs
#> 98  0.12547087 0.4768638 0.07080695 0.2799023 0.04695616 0.086371831 cult_trans
#> 99  0.15284634 0.3868815 0.07108708 0.3517291 0.03745606 0.140525102   fam_jobs
#> 100 0.17304831 0.3706235 0.08365837 0.3290363 0.04363356 0.027868104  fam_trans
#> 101 0.16802835 0.4589418 0.07245375 0.2569019 0.04367414 0.040489841 cult_house
#> 102 0.17719095 0.4265253 0.06865103 0.2672261 0.06040660 0.214584332  fam_trans
#> 103 0.17516616 0.4949700 0.07163167 0.2235549 0.03467731 0.044968979  fam_trans
#> 104 0.11301635 0.4778626 0.06799438 0.2994124 0.04171430 0.062997297  cult_jobs
#> 105 0.15836199 0.3896299 0.08258175 0.3253131 0.04411321 0.030668122  fam_trans
#> 106 0.16318526 0.3664394 0.08784100 0.3309043 0.05163010 0.152704219  cult_jobs
#> 107 0.15110057 0.5509970 0.06810769 0.1965247 0.03326996 0.061978192   cult_fam
#> 108 0.14003015 0.4810110 0.07630569 0.2526515 0.05000166 0.102864005  fam_trans
#> 109 0.11330634 0.4881247 0.06680591 0.2855801 0.04618291 0.140165677  cult_jobs
#> 110 0.12975723 0.5303311 0.06806407 0.2418788 0.02996888 0.089837749   fam_jobs
#> 111 0.18453730 0.3815080 0.07548072 0.3230647 0.03540923 0.030392370  fam_trans
#> 112 0.13942198 0.4969776 0.06317105 0.2722822 0.02814717 0.138176938   cult_fam
#> 113 0.14919701 0.3786648 0.07866926 0.3578431 0.03562586 0.141249309   fam_jobs
#> 114 0.12499633 0.4817004 0.08814703 0.2642621 0.04089421 0.100190603  fam_trans
#> 115 0.14564708 0.5161426 0.06237793 0.2377260 0.03810644 0.066011018   cult_fam
#> 116 0.10697763 0.5003206 0.07301107 0.2763642 0.04332646 0.049769588  fam_trans
#> 117 0.15626417 0.4481805 0.11046494 0.2451570 0.03993335 0.113028067  fam_house
#> 118 0.15007836 0.4825928 0.06808205 0.2524364 0.04681045 0.096285868  fam_trans
#> 119 0.13969265 0.4520390 0.07596160 0.2980188 0.03428790 0.023598140  cult_jobs
#> 120 0.16869015 0.4368763 0.05678742 0.3074338 0.03021225 0.203951002   fam_jobs
#> 121 0.13826153 0.4842814 0.06029813 0.2804513 0.03670766 0.102585844  cult_jobs
#> 122 0.21564683 0.3943112 0.07644977 0.2796303 0.03396194 0.204550024   fam_jobs
#> 123 0.15788036 0.3055999 0.09062489 0.3989455 0.04694940 0.081104221  cult_jobs
#> 124 0.16882138 0.4100042 0.08031520 0.2849412 0.05591796 0.091503408  fam_trans
#> 125 0.15458458 0.5556818 0.06869314 0.1853575 0.03568301 0.093269855   cult_fam
#> 126 0.11351876 0.5331797 0.06286350 0.2605098 0.02992823 0.069839141  cult_jobs
#> 127 0.17998377 0.3825615 0.07619791 0.3224089 0.03884787 0.020727243   fam_jobs
#> 128 0.26192257 0.3706408 0.09437494 0.2277276 0.04533413 0.127905282  fam_trans
#> 129 0.16846915 0.4335768 0.08388274 0.2589947 0.05507659 0.109256095  fam_trans
#> 130 0.12111430 0.4691301 0.07474959 0.3009188 0.03408729 0.068719518   cult_fam
#> 131 0.13301362 0.4027188 0.07442743 0.3493066 0.04053355 0.085070912  cult_jobs
#> 132 0.19267670 0.4387871 0.08583800 0.2451945 0.03750376 0.065217202  fam_trans
#> 133 0.16468313 0.4092011 0.06942500 0.3148546 0.04183616 0.136100693  fam_trans
#> 134 0.24279589 0.3508495 0.08837675 0.2685733 0.04940461 0.077847351  fam_trans
#> 135 0.13251937 0.4624260 0.06709044 0.2996496 0.03831460 0.149039703   cult_fam
#> 136 0.17971819 0.3772604 0.07171644 0.3245197 0.04678532 0.054165766  fam_trans
#> 137 0.13569741 0.5216160 0.06435193 0.2407593 0.03757541 0.083963700   cult_fam
#> 138 0.18179550 0.4758890 0.08038042 0.2235275 0.03840764 0.078320927  fam_trans
#> 139 0.15521505 0.5257572 0.06964075 0.2098694 0.03951765 0.058875241   cult_fam
#> 140 0.17822235 0.4157316 0.06885403 0.3012735 0.03591857 0.043441284   fam_jobs
#> 141 0.10697078 0.4822206 0.06603971 0.3039599 0.04080894 0.051225578  cult_jobs
#> 142 0.12229987 0.3467689 0.06600599 0.4292186 0.03570666 0.117968695  cult_jobs
#> 143 0.15916745 0.4695050 0.07230309 0.2663180 0.03270639 0.069091380   cult_fam
#> 144 0.11865631 0.5132571 0.07625190 0.2415334 0.05030122 0.092829337  fam_trans
#> 145 0.15901029 0.5373657 0.06001809 0.2121360 0.03146991 0.075489686   cult_fam
#> 146 0.12020076 0.4755362 0.07735503 0.2799321 0.04697590 0.128741471  fam_trans
#> 147 0.18865894 0.4827879 0.06249619 0.2306729 0.03538405 0.074138920 cult_house
#> 148 0.21216755 0.3568612 0.08443361 0.2992226 0.04731497 0.056409399   cult_fam
#> 149 0.11176308 0.4334485 0.08231441 0.3319761 0.04049796 0.035514671 cult_trans
#> 150 0.12086924 0.4857646 0.07663808 0.2774071 0.03932099 0.031181248  cult_jobs
#> 151 0.11540356 0.5290393 0.06064759 0.2621591 0.03275044 0.142198316  cult_jobs
#> 152 0.17606447 0.4409015 0.07685893 0.2640305 0.04214454 0.071939979  fam_trans
#> 153 0.11002050 0.3441972 0.06767874 0.4440230 0.03408055 0.121915940   fam_jobs
#> 154 0.14246159 0.5019526 0.07437613 0.2427879 0.03842176 0.079885565  fam_trans
#> 155 0.15889195 0.3981636 0.07562681 0.3288851 0.03843260 0.007117328   fam_jobs
#> 156 0.14901734 0.3957539 0.06788973 0.3523801 0.03495894 0.125677233   fam_jobs
#> 157 0.20856299 0.4310521 0.08587245 0.2397341 0.03477833 0.120983633  cult_jobs
#> 158 0.15177326 0.4339707 0.05964413 0.3212862 0.03332574 0.076771591 cult_house
#> 159 0.14333570 0.4795892 0.06378285 0.2772520 0.03604030 0.191203442  cult_jobs
#> 160 0.14327965 0.4035288 0.07847334 0.3263296 0.04838856 0.084468776  cult_jobs
#> 161 0.19045191 0.3319471 0.06813145 0.3783862 0.03108338 0.060460902  cult_jobs
#> 162 0.14469118 0.4354299 0.06981088 0.3118431 0.03822500 0.068056947   cult_fam
#> 163 0.17032302 0.4460237 0.06680828 0.2750897 0.04175531 0.018715021  fam_trans
#> 164 0.16352349 0.4457794 0.05976226 0.2971988 0.03373611 0.070575291 cult_house
#> 165 0.15106476 0.4596312 0.05897083 0.2847967 0.04553653 0.114073223  fam_trans
#> 166 0.15837758 0.4494626 0.06900412 0.2864451 0.03671062 0.068581958  cult_jobs
#> 167 0.14442932 0.4662513 0.06268036 0.2909600 0.03567903 0.114354171  cult_jobs
#> 168 0.12450730 0.4565278 0.07327748 0.3073210 0.03836642 0.033638029  cult_jobs
#> 169 0.16104389 0.5069422 0.06956837 0.2131295 0.04931608 0.106907328  fam_trans
#> 170 0.13700871 0.4106564 0.07491922 0.3432131 0.03420254 0.082124824  cult_jobs
#> 171 0.15737670 0.5154159 0.06661213 0.2136572 0.04693813 0.065041399   cult_fam
#> 172 0.13797411 0.5147021 0.06442713 0.2407474 0.04214927 0.095311844  fam_trans
#> 173 0.09083563 0.5278869 0.06528718 0.2768628 0.03912747 0.127352181  cult_jobs
#> 174 0.13092828 0.5620661 0.06305757 0.2094832 0.03446484 0.062383927   cult_fam
#> 175 0.12094684 0.5127124 0.07974628 0.2406437 0.04595079 0.036060488   cult_fam
#> 176 0.19388659 0.4836385 0.06532622 0.2164096 0.04073916 0.079030701 cult_house
#> 177 0.14521016 0.4615813 0.06760530 0.2906569 0.03494636 0.108177368   cult_fam
#> 178 0.17685499 0.3081136 0.07286710 0.4056177 0.03654670 0.092514654   fam_jobs
#> 179 0.15472397 0.4456467 0.06514446 0.2903074 0.04417749 0.091923572   cult_fam
#> 180 0.11685641 0.5126918 0.07920403 0.2188987 0.07234912 0.246777618  fam_trans
#> 181 0.18590262 0.3602110 0.06792178 0.3505323 0.03543234 0.045655826  cult_jobs
#> 182 0.19298646 0.5175886 0.06118477 0.1853350 0.04290509 0.193581744  fam_trans
#> 183 0.17978391 0.4333725 0.06441642 0.2796494 0.04277771 0.041684100  fam_trans
#> 184 0.17910338 0.4808911 0.06888317 0.2306183 0.04050406 0.054918206  fam_trans
#> 185 0.14394742 0.5576810 0.06229978 0.2045296 0.03154221 0.069896336   cult_fam
#> 186 0.14483888 0.5329136 0.07592738 0.2067842 0.03953594 0.082702333   cult_fam
#> 187 0.13263352 0.3441709 0.06784459 0.4178726 0.03747846 0.093036914  cult_jobs
#> 188 0.12967779 0.3941653 0.05935695 0.3816060 0.03519398 0.174088313  cult_jobs
#> 189 0.13328866 0.4638348 0.07324710 0.2915719 0.03805762 0.048705520  cult_jobs
#> 190 0.12849217 0.3638045 0.06580959 0.4086244 0.03326932 0.075349801   fam_jobs
#> 191 0.14412499 0.3969724 0.08151767 0.3406121 0.03677284 0.057726951  fam_trans
#> 192 0.11646171 0.4856418 0.06328139 0.2916299 0.04298525 0.168387107  cult_jobs
#> 193 0.15040813 0.4748702 0.06760672 0.2712731 0.03584186 0.103865249  fam_trans
#> 194 0.23989034 0.4976637 0.05875398 0.1713905 0.03230150 0.153580644  cult_jobs
#> 195 0.12029958 0.3593445 0.07281180 0.4146590 0.03288515 0.069896228   fam_jobs
#> 196 0.16977454 0.5225621 0.06591020 0.2043822 0.03737091 0.057037055 cult_house
#> 197 0.17155972 0.5188118 0.05431128 0.2163617 0.03895546 0.057822722 cult_house
#> 198 0.12673992 0.4764195 0.07551931 0.2833354 0.03798580 0.042254453  cult_jobs
#> 199 0.14982086 0.5501472 0.05281082 0.2060319 0.04118916 0.123320565  fam_trans
#> 200 0.12527603 0.4510233 0.06668400 0.3243587 0.03265796 0.062652221   cult_fam
#>           top2        top3
#> 1    fam_trans  house_jobs
#> 2    fam_trans   cult_jobs
#> 3   cult_trans    fam_jobs
#> 4    cult_jobs  cult_house
#> 5    fam_trans    fam_jobs
#> 6     cult_fam  cult_house
#> 7     cult_fam  cult_house
#> 8   cult_house    cult_fam
#> 9   cult_house  house_jobs
#> 10    cult_fam   cult_jobs
#> 11    cult_fam  cult_house
#> 12    cult_fam  cult_house
#> 13  cult_house   fam_house
#> 14   fam_trans  cult_house
#> 15  cult_house   fam_trans
#> 16  cult_trans    fam_jobs
#> 17   fam_trans   cult_jobs
#> 18    cult_fam   fam_trans
#> 19   cult_jobs  cult_house
#> 20    cult_fam   fam_house
#> 21  cult_house   fam_trans
#> 22    cult_fam  cult_house
#> 23    cult_fam   fam_trans
#> 24    cult_fam  cult_house
#> 25    cult_fam house_trans
#> 26    cult_fam   fam_house
#> 27    cult_fam    fam_jobs
#> 28   fam_trans  cult_house
#> 29    fam_jobs   cult_jobs
#> 30    cult_fam   fam_trans
#> 31    cult_fam  cult_house
#> 32   fam_trans  house_jobs
#> 33    cult_fam   cult_jobs
#> 34   fam_trans   cult_jobs
#> 35  cult_house   cult_jobs
#> 36    fam_jobs   fam_trans
#> 37   fam_trans   cult_jobs
#> 38   fam_trans    fam_jobs
#> 39  cult_house    fam_jobs
#> 40  cult_trans  cult_house
#> 41   fam_trans   fam_house
#> 42    cult_fam  cult_house
#> 43    cult_fam  house_jobs
#> 44   cult_jobs    cult_fam
#> 45    cult_fam   cult_jobs
#> 46    cult_fam  cult_house
#> 47    fam_jobs  cult_house
#> 48    fam_jobs  cult_house
#> 49    cult_fam  cult_house
#> 50  cult_house   fam_trans
#> 51  house_jobs    fam_jobs
#> 52   cult_jobs   fam_house
#> 53   fam_trans    cult_fam
#> 54  cult_house   fam_trans
#> 55   cult_jobs    cult_fam
#> 56   cult_jobs   fam_trans
#> 57   cult_jobs    fam_jobs
#> 58  cult_house    cult_fam
#> 59  cult_house   fam_trans
#> 60   cult_jobs    cult_fam
#> 61   cult_jobs   fam_house
#> 62    cult_fam  cult_house
#> 63    cult_fam  cult_house
#> 64   fam_trans  cult_trans
#> 65  cult_house   fam_trans
#> 66    cult_fam   fam_house
#> 67    fam_jobs   cult_jobs
#> 68   fam_trans  cult_house
#> 69    cult_fam   cult_jobs
#> 70   fam_trans    fam_jobs
#> 71   cult_jobs    cult_fam
#> 72   fam_house  cult_trans
#> 73   fam_trans  cult_house
#> 74    fam_jobs    cult_fam
#> 75   fam_trans   cult_jobs
#> 76    cult_fam    fam_jobs
#> 77   fam_trans    cult_fam
#> 78  cult_house   fam_trans
#> 79    cult_fam  cult_house
#> 80   fam_trans    cult_fam
#> 81    fam_jobs   fam_trans
#> 82   fam_trans    cult_fam
#> 83   fam_trans  house_jobs
#> 84  cult_house   fam_trans
#> 85    cult_fam   fam_trans
#> 86    cult_fam   cult_jobs
#> 87   cult_jobs  cult_house
#> 88   cult_jobs    cult_fam
#> 89    cult_fam   fam_house
#> 90   fam_house  cult_house
#> 91    fam_jobs   fam_trans
#> 92    cult_fam   fam_house
#> 93   fam_trans    fam_jobs
#> 94    fam_jobs  cult_house
#> 95   fam_trans   cult_jobs
#> 96   fam_trans    fam_jobs
#> 97   fam_trans    fam_jobs
#> 98   fam_trans    cult_fam
#> 99   fam_trans    cult_fam
#> 100   fam_jobs  cult_house
#> 101 cult_trans   fam_trans
#> 102 cult_house  house_jobs
#> 103   fam_jobs  cult_house
#> 104   cult_fam   fam_trans
#> 105   fam_jobs  cult_trans
#> 106  fam_trans    cult_fam
#> 107   fam_jobs  cult_house
#> 108   cult_fam  cult_house
#> 109  fam_trans    cult_fam
#> 110  cult_jobs    cult_fam
#> 111   fam_jobs  cult_house
#> 112  cult_jobs  cult_house
#> 113   cult_fam   fam_trans
#> 114   cult_fam   cult_jobs
#> 115 house_jobs   fam_trans
#> 116   cult_fam   cult_jobs
#> 117  fam_trans    cult_fam
#> 118  cult_jobs  cult_house
#> 119  fam_trans  house_jobs
#> 120   cult_fam  cult_house
#> 121  fam_trans  cult_house
#> 122  cult_jobs    cult_fam
#> 123  fam_trans    fam_jobs
#> 124 cult_house  house_jobs
#> 125   fam_jobs  cult_house
#> 126   cult_fam   fam_house
#> 127 cult_house   fam_trans
#> 128  cult_jobs    cult_fam
#> 129   cult_fam   cult_jobs
#> 130  cult_jobs    fam_jobs
#> 131  fam_trans   fam_house
#> 132   fam_jobs  cult_trans
#> 133  cult_jobs    fam_jobs
#> 134 cult_house   cult_jobs
#> 135  fam_trans   cult_jobs
#> 136 cult_house    fam_jobs
#> 137 cult_house   fam_trans
#> 138   cult_fam house_trans
#> 139  fam_trans  jobs_trans
#> 140  fam_trans   fam_house
#> 141 cult_trans    cult_fam
#> 142   fam_jobs   fam_trans
#> 143   fam_jobs   fam_trans
#> 144   cult_fam  cult_trans
#> 145  fam_trans   fam_house
#> 146   cult_fam   cult_jobs
#> 147  fam_trans    cult_fam
#> 148  fam_trans  cult_trans
#> 149  cult_jobs    fam_jobs
#> 150   cult_fam   fam_trans
#> 151   cult_fam    fam_jobs
#> 152 cult_house    cult_fam
#> 153  cult_jobs    cult_fam
#> 154   cult_fam    fam_jobs
#> 155  fam_trans   cult_jobs
#> 156   cult_fam  cult_house
#> 157  fam_trans    cult_fam
#> 158  cult_jobs    cult_fam
#> 159  fam_trans    fam_jobs
#> 160 cult_trans   fam_trans
#> 161 cult_house    fam_jobs
#> 162 cult_house    fam_jobs
#> 163 cult_house   cult_jobs
#> 164   cult_fam    fam_jobs
#> 165   cult_fam   cult_jobs
#> 166 cult_house    cult_fam
#> 167  fam_trans    cult_fam
#> 168  fam_trans  house_jobs
#> 169   cult_fam   fam_house
#> 170  fam_trans    cult_fam
#> 171  fam_trans  cult_house
#> 172   cult_fam  house_jobs
#> 173  fam_trans    cult_fam
#> 174   fam_jobs   fam_trans
#> 175 jobs_trans   fam_trans
#> 176  fam_trans    cult_fam
#> 177  fam_trans    fam_jobs
#> 178  fam_trans  cult_house
#> 179  fam_trans    fam_jobs
#> 180   fam_jobs    cult_fam
#> 181 cult_house   fam_trans
#> 182   cult_fam   cult_jobs
#> 183   fam_jobs  cult_house
#> 184  fam_house  cult_house
#> 185 cult_house    fam_jobs
#> 186  fam_trans    fam_jobs
#> 187  fam_trans    fam_jobs
#> 188  fam_trans    cult_fam
#> 189  fam_trans    cult_fam
#> 190  cult_jobs   fam_trans
#> 191  cult_jobs    cult_fam
#> 192  fam_trans    cult_fam
#> 193  cult_jobs  cult_house
#> 194   cult_fam   fam_trans
#> 195  cult_jobs   fam_trans
#> 196   cult_fam   fam_trans
#> 197   cult_fam   fam_trans
#> 198  fam_trans    cult_fam
#> 199   cult_fam    fam_jobs
#> 200  cult_jobs    fam_jobs