Aggregate individual judgments

Aggregate individual judgments from pairwise comparison matrices

Usage

ahp.aggjudge(ahpmat, atts, aggmethod = "geometric", qt = 0)

Arguments

ahpmat

A list of pairwise comparison matrices of each decision maker generated by ahp.mat.

atts

a list of attributes in the correct order

aggmethod

The method of aggregating the judgments by all decision-makers. Five modes aggregation are available: arithmetic: the arithmetic mean; geometric: the geometric mean (the default); rootmean: the square root of the sum of the squared value, tmean: the trimmed mean, tgmean: trimmed geometric mean. The quantiles trimmed are based on qt. It can also be set to sd, where it reports the standard deviation from the arithmetic mean.

qt

specifies the quantile which the top and bottom priority weights are trimmed. Used only if aggmethod = 'tmean' or aggmethod = 'tgmean'. For example, qt = 0.25 specifies that the aggregation is the arithmetic mean of the values from the 25 to 75 percentile. By default qt = 0.

Value

A data.frame of the aggregated pairwise judgments of all the decision-makers.

Author

Frankie Cho

Examples

## Computes individual judgments with geometric mean and aggregates them
## with a trimmed arithmetic mean

data(city200)
atts <- c('cult', 'fam', 'house', 'jobs', 'trans')

cityahp <- ahp.mat(df = city200, atts = atts, negconvert = TRUE)
ahp.aggjudge(cityahp, atts, aggmethod = 'tmean', qt = 0.1)
#>            cult       fam    house      jobs    trans
#> cult        NaN 0.2287760 3.000000 0.3975410 4.913669
#> fam   4.6484375       NaN 6.706667 1.6587302 6.525180
#> house 0.3333333 0.1525397      NaN       NaN      NaN
#> jobs  2.7049180 0.6706349      NaN       NaN 7.000000
#> trans 0.2091127 0.1576739      NaN 0.1428571      NaN