ahpsurvey: Analytic Hierarchy Process for Survey Data

The ahpsurvey package provides a workflow for researchers to reformat data and run analytic hierarchy process on data that are originally formatted using the survey data entry mode. A diversified toolbox is available for researchers to explore ways to impute missing pairwise comparisons or to adjust comparisons based on consistency, and to conveniently customize the methods which weights derived from surveys are extracted and aggregated.

Author

Frankie Cho frankie.cho@monash.edu