A dataframe of 200 simulated decision-makers based on Saaty (2004)

A data.frame of one decision-maker with weights the same as Saaty (2004) , with the pairwise comparisons of atts: c('cult', 'fam', 'house', 'jobs', 'trans'), about a choosing the best city to live in based on five attributes: Culture, Family, House, Jobs, and Transportation. The choices of 200 decision-makers are simulated based on the underlying weights and randomness generated by a normal random distribution. The true weights are stored in data(city1). The normal random distribution was set with a mean at the true weight value and a standard deviation of 2,1,2,1.5,2,1.5,2.5,0.5,0.5,1 in chronological order of the columns in city200.

Usage

data(city200)

Format

A data.frame with 200 row and 10 variables, which are pairwise comparisons of atts.

Source

Saaty TL (2004). “Decision making — the Analytic Hierarchy and Network Processes (AHP/ANP).” Journal of Systems Science and Systems Engineering, 13(1), 1–35. ISSN 1861-9576. doi:10.1007/s11518-006-0151-5 .